Traffic Classifier Project
My first hand at Deep Learning. 
I've created my first neural network by means TensorFlow in order to correctly detect Traffic Signs. 
The net was trained over a given dataset, which was splitted also in a validation set so that evaluate the accuracy of the model.
Further in the last part is shown how the model react at individuating the same type of signs, but with the difference that these new signs   have been retrieved on the web and they didn't belong to the original dataset.